+++ Src/Render/PlanetRendererMed.cpp
@@ -26,7 +26,7 @@ void MedDetailPlanetRenderer::Update(float dt)
 {
     dt;
 
-    m_world *= Matrix::CreateScale(m_planet->GetRadius() / Constants::Scale);
+    m_world *= Matrix::CreateScale((float)(m_planet->GetRadius() / Constants::Scale));
 }
 
 void Galactic::MedDetailPlanetRenderer::Reset()
